remove warning

+++ src/search/hddehc.cc
@@ -110,7 +110,7 @@ int HDDEHC::Search() {
   auto state = InitialEvaluate();
 
   if (runup() && rank() == initial_rank()) {
-    while (n_open_nodes() < world_size() && !NoNode()) {
+    while (static_cast<int>(n_open_nodes()) < world_size() && !NoNode()) {
       int node = Pop();
       int goal = IndependentExpand(node, state);
