+++ src/search/symmetry_breaking_hdgbfs.cc
@@ -139,7 +139,7 @@ int SBHDGBFS::Search() {
   auto state = InitialEvaluate();
 
   if (runup() && rank() == initial_rank()) {
-    while (n_open_nodes() < world_size() && !NoNode()) {
+    while (static_cast<int>(n_open_nodes()) < world_size() && !NoNode()) {
       int node = Pop();
       int goal = IndependentExpand(node, state);
